Add modal for voters and create filter
