+++ packages/df-blogs/src/BlogFollowers.tsx
@@ -1,0 +1,59 @@
+import React, { useState } from 'react';
+
+import { withCalls, withMulti } from '@polkadot/ui-api/with';
+import { AccountId } from '@polkadot/types';
+import { queryBlogsToProp } from './utils';
+import { Modal, Button } from 'semantic-ui-react';
+import _ from 'lodash';
+import AddressMini from '@polkadot/ui-app/AddressMiniJoy';
+
+type Props = {
+  followers?: AccountId[],
+  followersCount: Number
+};
+
+const InnerBlogFollowersModal = (props: Props) => {
+
+  const { followers, followersCount } = props;
+  const [open, setOpen] = useState(false);
+
+  const renderFollowers = () => {
+    return followers && followers.map(account =>
+      <div style={{ textAlign: 'left', margin: '1rem' }}>
+        <AddressMini
+          value={account}
+          isShort={false}
+          isPadded={false}
+          size={48}
+          withName
+          withBalance
+        />
+      </div>
+    );
+  };
+
+  return (
+    <Modal
+      open={open}
+      dimmer='blurring'
+      trigger={<Button basic onClick={() => setOpen(true)}>Followers ({followersCount})</Button>}
+      centered={true}
+      style={{ marginTop: '3rem' }}
+    >
+      <Modal.Header><h1>Blog followers ({followersCount})</h1></Modal.Header>
+      <Modal.Content scrolling>
+        {renderFollowers()}
+      </Modal.Content>
+      <Modal.Actions>
+        <Button content='Close' onClick={() => setOpen(false)} />
+      </Modal.Actions>
+    </Modal>
+  );
+};
+
+export const BlogFollowersModal = withMulti(
+  InnerBlogFollowersModal,
+  withCalls<Props>(
+    queryBlogsToProp('blogFollowers', { paramName: 'id', propName: 'followers' })
+  )
+);
