+++ packages/df-blogs/src/ViewModalVoters.tsx
@@ -1,64 +1,13 @@
 import React, { useState, useEffect } from 'react';
 
 import { withCalls, withMulti } from '@polkadot/ui-api/with';
-import { AccountId, Option } from '@polkadot/types';
 import { queryBlogsToProp } from './utils';
-import { Modal, Button } from 'semantic-ui-react';
+import { Modal, Button, Tab } from 'semantic-ui-react';
 import _ from 'lodash';
 import AddressMini from '@polkadot/ui-app/AddressMiniJoy';
 import { ReactionId, Post, Reaction } from './types';
 import { api } from '@polkadot/ui-api/Api';
 
-type Props = {
-  followers?: AccountId[],
-  followersCount: Number
-};
-
-const InnerBlogFollowersModal = (props: Props) => {
-
-  const { followers, followersCount } = props;
-  const [open, setOpen] = useState(false);
-
-  const renderFollowers = () => {
-    return followers && followers.map(account =>
-      <div style={{ textAlign: 'left', margin: '1rem' }}>
-        <AddressMini
-          value={account}
-          isShort={false}
-          isPadded={false}
-          size={48}
-          withName
-          withBalance
-        />
-      </div>
-    );
-  };
-
-  return (
-    <Modal
-      open={open}
-      dimmer='blurring'
-      trigger={<Button basic onClick={() => setOpen(true)}>Followers ({followersCount})</Button>}
-      centered={true}
-      style={{ marginTop: '3rem' }}
-    >
-      <Modal.Header><h1>Blog followers ({followersCount})</h1></Modal.Header>
-      <Modal.Content scrolling>
-        {renderFollowers()}
-      </Modal.Content>
-      <Modal.Actions>
-        <Button content='Close' onClick={() => setOpen(false)} />
-      </Modal.Actions>
-    </Modal>
-  );
-};
-
-export const BlogFollowersModal = withMulti(
-  InnerBlogFollowersModal,
-  withCalls<Props>(
-    queryBlogsToProp('blogFollowers', { paramName: 'id', propName: 'followers' })
-  )
-);
 
 type PropsVoters = {
   id: Comment | Post,
@@ -87,20 +36,29 @@ const InnerVotersModal = (props: PropsVoters) => {
     loadVoters();
   });
 
-  const renderVoters = () => {
-    return reactionView.map(reaction => {
-      return <div style={{ textAlign: 'left', margin: '1rem' }}>
+  const renderVoters = (state: Array<Reaction>) => {
+    return state.map(reaction => {
+      return <div key={reaction.id.toNumber()} style={{ textAlign: 'left', margin: '1rem' }}>
       <AddressMini
         value={reaction.created.account}
-        isShort={true}
         isPadded={false}
         size={28}
-        withName
         extraDetails={`Kind: ${reaction.kind}`}
       />
     </div>;
     });
   };
+  const panes = [
+  { key: 'all', menuItem: 'All', render: () => <Tab.Pane>{renderVoters(reactionView)}</Tab.Pane> },
+    { key: 'upvote', menuItem: 'Upvoters', render: () => {
+      const reactionWithUpVoters = reactionView.filter(reaction => reaction.kind.toString() === 'Upvote');
+      return <Tab.Pane>{renderVoters(reactionWithUpVoters)}</Tab.Pane>;
+    }},
+    { key: 'downvote', menuItem: 'Downvoters', render: () => {
+      const reactionWithDownVoters = reactionView.filter(reaction => reaction.kind.toString() === 'Downvote');
+      return <Tab.Pane>{renderVoters(reactionWithDownVoters)}</Tab.Pane>;
+    }}
+  ];
 
   return (
     <Modal
@@ -111,7 +69,7 @@ const InnerVotersModal = (props: PropsVoters) => {
     >
       <Modal.Header><h1>Voters ({votersCount})</h1></Modal.Header>
       <Modal.Content scrolling>
-        {renderVoters()}
+      <Tab panes={panes} />
       </Modal.Content>
       <Modal.Actions>
         <Button content='Close' onClick={close} />
